server/requirements.txt
# FastAPI & Web Frameworks
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
python-decouple==3.8
email-validator==2.1.0
python-dateutil==2.8.2
typing-extensions==4.8.0

# Testing & Development Tools
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1 